+++ x509/src/ext/pkix/mod.rs
@@ -1,5 +1,6 @@
 //! PKIX X.509 Certificate Extensions (RFC 5280)
 
+pub mod certpolicy;
 pub mod name;
 pub mod oids;
 
@@ -7,6 +8,7 @@ mod authkeyid;
 mod keyusage;
 
 pub use authkeyid::AuthorityKeyIdentifier;
+pub use certpolicy::CertificatePolicies;
 pub use keyusage::{KeyUsage, KeyUsages};
 
 use der::asn1::OctetString;
